requirements.txt
# Telegram Bot依赖
python-telegram-bot==20.7

# HTTP请求库
aiohttp==3.9.1
requests==2.31.0

# 系统监控
psutil==5.9.6

# 数据处理
typing-extensions==4.8.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Web服务器（用于健康检查）
flask==3.0.0
waitress==2.1.2

# 日志和调试
coloredlogs==15.0.1

# 异步支持
asyncio-throttle==1.0.2

# 内存优化
memory-profiler==0.61.0

# 环境变量管理
python-dotenv==1.0.0

# JSON处理增强
orjson==3.9.10

# 时间处理
pytz==2023.3

# 错误处理和重试
tenacity==8.2.3

# 性能监控
prometheus-client==0.19.0

# 垃圾回收工具
gc-python-utils==1.0.0